SPARK-25860: Separate end-to-end suite and switch to tables
